+++ scripts/18-software/categories/multimedia.sh
@@ -2,6 +2,9 @@
 
 echo "   - Multimedia..."
 
+LOCK=/tmp/workstation.lock.multimedia
+[ -f ${LOCK} ] && return
+
 # From repository
 apt-get install -y audacity       \
                    handbrake      \
@@ -10,3 +13,5 @@ apt-get install -y audacity       \
                    spotify-client \
 > /dev/null 2>&1
 # From repository
+
+> ${LOCK}
